# docs/spec/openapi.json
{"openapi":"3.1.0","info":{"title":"app","version":"version"},"paths":{"/":{"get":{"operationId":"global.index","tags":["Default"]}},"/status":{"get":{"operationId":"global.status","tags":["Default"]}},"/spec/openapi.json":{"get":{"operationId":"global.spec/openapi.json","tags":["Default"]}},"/spec":{"get":{"operationId":"global.spec","tags":["Default"]}},"/adminme":{"get":{"operationId":"global.adminme","tags":["Default"]}},"/auth/{provider}/redirect":{"get":{"operationId":"AuthController.redirect","tags":["Auth"],"description":"Redirects to the social provider for authentication","responses":{"307":{"description":"Redirects to the social provider"},"422":{"description":"Provided request data did not pass validation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationError"}}}}},"parameters":[{"name":"provider","in":"path","required":true,"schema":{"enum":["discord","github"]}}]}},"/auth/{provider}/callback":{"get":{"operationId":"AuthController.callback","tags":["Auth"],"description":"Handles the callback from the social provider","responses":{"307":{"description":"Redirects to the status page"},"422":{"description":"Provided request data did not pass validation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationError"}}}}},"parameters":[{"name":"provider","in":"path","required":true,"schema":{"enum":["discord","github"]}}]}},"/auth/register":{"post":{"operationId":"AuthController.register","tags":["Auth"],"description":"Registers a new user with email and password","responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Provided request data did not pass validation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationError"}}}}},"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"nickname":{"type":"string","minLength":3,"maxLength":16},"name":{"type":"string","minLength":3,"maxLength":32},"surname":{"type":"string","minLength":3,"maxLength":32},"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":8}},"required":["nickname","email","password"],"additionalProperties":false}}}}}},"/auth/login":{"post":{"operationId":"AuthController.login","tags":["Auth"],"description":"Logs in a user with email and password","responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Provided request data did not pass validation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationError"}}}}},"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["email","password"],"additionalProperties":false}}}}}},"/auth/logout":{"post":{"operationId":"AuthController.logout","tags":["Auth"],"description":"Logs out the current user","responses":{"200":{"description":"User logged out successfully"}}}},"/events":{"get":{"operationId":"EventsController.index","tags":["Events"]},"post":{"operationId":"EventsController.store","tags":["Events"]}},"/events/{id}":{"get":{"operationId":"EventsController.show","tags":["Events"]},"put":{"operationId":"EventsController.update","tags":["Events"]},"patch":{"operationId":"EventsController.update","tags":["Events"]},"delete":{"operationId":"EventsController.destroy","tags":["Events"]}},"/events/{id}/administrators":{"get":{"operationId":"EventsController.indexAdministrators","tags":["Events"]},"post":{"operationId":"EventsController.storeAdministrator","tags":["Events"]}},"/events/{id}/administrators/{adminId}":{"put":{"operationId":"EventsController.updateAdministrator","tags":["Events"]},"patch":{"operationId":"EventsController.updateAdministrator","tags":["Events"]},"delete":{"operationId":"EventsController.destroyAdministrator","tags":["Events"]}},"/teams/{id}":{"get":{"operationId":"TeamsController.show","tags":["Teams"]},"put":{"operationId":"TeamsController.update","tags":["Teams"]},"patch":{"operationId":"TeamsController.update","tags":["Teams"]},"delete":{"operationId":"TeamsController.destroy","tags":["Teams"]}},"/events/{event}_id/teams":{"post":{"operationId":"TeamsController.store","tags":["Teams"]},"get":{"operationId":"TeamsController.index","tags":["Teams"]}},"/teams/{id}/invites":{"post":{"operationId":"TeamsController.storeInvite","tags":["Teams"]},"get":{"operationId":"TeamsController.indexInvites","tags":["Teams"]}},"/teams/{id}/kick":{"post":{"operationId":"TeamsController.kickMember","tags":["Teams"]}},"/invitations/{id}":{"get":{"operationId":"TeamsController.respondToInvite","tags":["Teams"]}},"/invitations":{"get":{"operationId":"TeamsController.indexUserInvites","tags":["Teams"]}},"/events/{event}_id/organizations":{"get":{"operationId":"OrganizationsController.index","tags":["Organizations"]},"post":{"operationId":"OrganizationsController.store","tags":["Organizations"]}},"/events/{event}_id/organizations/{id}":{"get":{"operationId":"OrganizationsController.show","tags":["Organizations"]},"put":{"operationId":"OrganizationsController.update","tags":["Organizations"]},"patch":{"operationId":"OrganizationsController.update","tags":["Organizations"]},"delete":{"operationId":"OrganizationsController.destroy","tags":["Organizations"]}},"/tasks/{id}/sponsors":{"get":{"operationId":"OrganizationsController.indexSponsors","tags":["Organizations"]},"post":{"operationId":"OrganizationsController.storeSponsor","tags":["Organizations"]}},"/tasks/{id}/sponsors/{sponsorId}":{"get":{"operationId":"OrganizationsController.showSponsor","tags":["Organizations"]},"delete":{"operationId":"OrganizationsController.destroySponsor","tags":["Organizations"]}},"/tasks/{id}":{"get":{"operationId":"TasksController.show","tags":["Tasks"]},"put":{"operationId":"TasksController.update","tags":["Tasks"]},"patch":{"operationId":"TasksController.update","tags":["Tasks"]},"delete":{"operationId":"TasksController.destroy","tags":["Tasks"]}},"/tasks/{taskId}/registrations":{"post":{"operationId":"TasksController.storeTaskRegistration","tags":["Tasks"]}},"/registrations/{id}":{"delete":{"operationId":"TasksController.destroyTaskRegistration","tags":["Tasks"]}},"/hackathon/tasks/{id}":{"put":{"operationId":"HackathonController.updateTask","tags":["Hackathons"]},"patch":{"operationId":"HackathonController.updateTask","tags":["Hackathons"]}},"/hackathon/tasks/{id}/jury":{"get":{"operationId":"HackathonController.indexJuryMembers","tags":["Hackathons"]},"post":{"operationId":"HackathonController.storeJuryMember","tags":["Hackathons"]}},"/hackathon/tasks/{id}/jury/{juryMemberId}":{"get":{"operationId":"HackathonController.showJuryMember","tags":["Hackathons"]},"put":{"operationId":"HackathonController.updateJuryMember","tags":["Hackathons"]},"patch":{"operationId":"HackathonController.updateJuryMember","tags":["Hackathons"]},"delete":{"operationId":"HackathonController.destroyJuryMember","tags":["Hackathons"]}},"/event/{event}_id/tasks":{"get":{"operationId":"TasksController.index","tags":["Tasks"]}},"/event/{event}_id/task":{"post":{"operationId":"TasksController.store","tags":["Tasks"]}},"/tasks/{task}_id/scores":{"get":{"operationId":"ScoresController.indexCriteria","tags":["Scores"]},"post":{"operationId":"ScoresController.storeCriteria","tags":["Scores"]}},"/tasks/{task}_id/scores/{id}":{"get":{"operationId":"ScoresController.showCriteria","tags":["Scores"]},"put":{"operationId":"ScoresController.updateCriteria","tags":["Scores"]},"patch":{"operationId":"ScoresController.updateCriteria","tags":["Scores"]},"delete":{"operationId":"ScoresController.destroyCriteria","tags":["Scores"]}}},"components":{"schemas":{"ValidationError":{"type":"object","properties":{"errors":{"type":"array","items":{"type":"object","properties":{"message":{"type":"string"},"rule":{"type":"string"},"field":{"type":"string"}},"required":["message","rule","field"],"additionalProperties":false}}},"required":["errors"],"additionalProperties":false},"User":{"type":"object","properties":{"id":{"type":"number"},"nickname":{"example":"JohnnyBravo","type":"string"},"name":{"example":"John","type":"string"},"surname":{"example":"Bravo","type":"string"},"email":{"example":"johnny.bravo@example.com","type":"string"},"avatarUrl":{"example":"https://example.com/avatar.png","type":"string"},"permissions":{"format":"binary","type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","nickname","email","permissions","createdAt","updatedAt"]}},"responses":{},"parameters":{},"examples":{},"requestBodies":{},"headers":{},"securitySchemes":{},"links":{},"callbacks":{}},"tags":[{"name":"Default","description":"Default tag for all non-classified endpoints"},{"name":"Auth","description":"Endpoints provided by AuthController"},{"name":"Events","description":"Endpoints provided by EventsController"},{"name":"Teams","description":"Endpoints provided by TeamsController"},{"name":"Organizations","description":"Endpoints provided by OrganizationsController"},{"name":"Tasks","description":"Endpoints provided by TasksController"},{"name":"Hackathons","description":"Endpoints provided by HackathonsController"},{"name":"Scores","description":"Endpoints provided by ScoresController"}],"servers":[]}